feat: Refactor MessageService and DTOs for ConversationMedia integration

+++ api/src/main/java/com/ruby/rubia_server/core/dto/CreateMessageDTO.java
@@ -44,5 +44,5 @@ public class CreateMessageDTO {
     @Builder.Default
     private Boolean isAiGenerated = false;
     
-    private BigDecimal aiConfidence;
+    private Double aiConfidence;
 }
